Handle single file input, import line deletion and clear multi-line rewrites

+++ main.py
@@ -1,20 +1,35 @@
-import ast
 import os
-import re
 import sys
-from itertools import count
 
 from transformer import Transformer
 
-for root, dirs, files in os.walk(sys.argv[1]):
-    for file in files:
-        if os.path.splitext(file)[1] == ".py":
-            full_path = os.path.join(root, file)
-            print(f"Rewriting {full_path}")
-            with open(full_path) as f:
-                source = f.read()
-            transformer = Transformer()
-            source = transformer.transform(source)
-            re.sub(r"(\s+\\$)+\s+", "", source)
-            with open(full_path, "w") as f:
-                f.write(source)
+path = sys.argv[1]
+
+if not os.path.exists(path):
+    raise ValueError(f"{path} not found")
+
+
+def rewrite_file(full_path):
+    print(f"Rewriting {full_path}")
+    with open(full_path) as f:
+        source = f.read()
+    transformer = Transformer()
+    source = transformer.transform(source)
+    with open(full_path, "w") as f:
+        f.write(source)
+
+
+if os.path.isfile(path):
+    rewrite_file(path)
+else:
+    for root, dirs, files in os.walk(path):
+        for file in files:
+            if os.path.splitext(file)[1] == ".py":
+                full_path = os.path.join(root, file)
+                print(f"Rewriting {full_path}")
+                with open(full_path) as f:
+                    source = f.read()
+                transformer = Transformer()
+                source = transformer.transform(source)
+                with open(full_path, "w") as f:
+                    f.write(source)
